functions/saving_job_failures_bq/requirements.txt
functions-framework==3.3.0
google-cloud-bigquery==3.5.0